# projects/functions/axlr8/src/openapi.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"postDataString": {
				"properties": {
					"rptId": {
						"type": "number",
						"format": "double"
					},
					"rptFilters": {
						"items": {
							"properties": {
								"values": {
									"type": "string"
								},
								"operator": {
									"type": "string"
								},
								"filter": {
									"type": "string"
								}
							},
							"required": [
								"values",
								"operator",
								"filter"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"rptId",
					"rptFilters"
				],
				"type": "object",
				"additionalProperties": false
			},
			"reportResult": {
				"properties": {
					"result": {
						"type": "boolean"
					},
					"success": {
						"anyOf": [
							{
								"items": {
									"properties": {},
									"additionalProperties": {
										"anyOf": [
											{
												"type": "string"
											},
											{
												"type": "number",
												"format": "double"
											}
										]
									},
									"type": "object"
								},
								"type": "array"
							},
							{
								"type": "boolean",
								"enum": [
									false
								]
							}
						]
					},
					"error": {
						"type": "string",
						"nullable": true
					},
					"debug": {
						"$ref": "#/components/schemas/postDataString"
					}
				},
				"required": [
					"result",
					"success",
					"error",
					"debug"
				],
				"type": "object",
				"additionalProperties": false
			},
			"report200": {
				"enum": [
					"OSCE_APPLIED_DATE",
					"OSCE_CONT_DOB",
					"OSCE_CONT_EMAIL",
					"OSCE_CONTACT_ID",
					"OSCE_CONTACT_ID",
					"OSCE_PROMOTE_DATE",
					"OSCE_EMP_RETIRED_DATE",
					"OSCE_CONT_1STNAME",
					"OSCE_CONT_LASTNAME",
					"OSCE_FULL_NAME",
					"OSCE_CONT_SEX",
					"OSCE_VETTING_INTERVIEWED_BY",
					"ES_JOB_APP_STATUS",
					"OSCE_JAS_LAST_CHANGE",
					"OSCE_CONTACT_MOBILE",
					"OSCE_EMP_NI",
					"OSCE_ORG_SOURCE",
					"OSCE_ORG_TYPE",
					"ES_EMP_PAYROLL_NO",
					"OSCE_SITE_POSTCODE",
					"SECT_QUAL_NAME",
					"SECT_QUAL_PASS_DATE",
					"SECT_QUAL_TO",
					"OSCE_MOST_RECENT_BOOKING_DATE",
					"OSCE_EMP_PROMOTED_BY",
					"OSCE_EMP_APPLICANT_MANAGER",
					"SECT_QUAL_CERTIFICATE",
					"OSCE_EMP_ACTIVE",
					"OSCE_JP_JAS_HISTORY_JPJ_DATE_CHANGED",
					"OSCE_JP_JAS_HISTORY_JPJ_CHANGEUSER",
					"OSCE_JP_JAS_HISTORY_JPJ_ID",
					"OSCE_JP_JAS_HISTORY_JPJ_CES_STATUS"
				],
				"type": "string"
			},
			"report208": {
				"enum": [
					"OSC_ORG_ALPHA",
					"OSC_ORG_ACC_MGR",
					"OSC_ORG_NAME",
					"LU_LKH_ORG_410",
					"LU_LKH_SITE_411"
				],
				"type": "string"
			},
			"report222": {
				"enum": [
					"OSC_ORG_ALPHA",
					"ES_ORG_ALPHA",
					"ES_EVENT_NAME_SELECT",
					"ES_EVENT_NAME_SELECT",
					"ES_ID",
					"ES_SITE_CUSTOMER_ID",
					"ES_SESSION_NUMBER_REQ",
					"ES_SESSION_NUMBER_STAFF_STILL_REQUIRED",
					"ES_NUMBER_DNS",
					"ES_SESSION_START_DATE",
					"ES_ROLE_START_TIME",
					"ES_ROLE_END_TIME",
					"ES_BOOKING_KEY",
					"EMP_BOOKING_STATUS",
					"ES_NO_CONFIRMED",
					"ES_BOOKING_WORKED_STATUS",
					"EMP_BOOKING_AVAILABILITY_STATUS",
					"EMP_DECLINED_STATUS",
					"ES_DID_NOT_SHOW",
					"ES_BOOKING_DNS_REASON",
					"ES_SESSION_BOOKING_HOUR_RATE",
					"ES_SESSION_BOOKING_DAY_RATE",
					"ES_RATE",
					"ES_DAY_RATE",
					"ES_SESSION_ROLE_RATE",
					"ES_CONTACT_FULLNAME",
					"LU_LKH_CONTACTS_45",
					"ES_CONTACT_EMAIL",
					"ES_EMP_APPLICANT_MANAGER",
					"ES_CONT_STRIKES"
				],
				"type": "string"
			},
			"stelify": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/report200"
					},
					{
						"$ref": "#/components/schemas/report208"
					},
					{
						"$ref": "#/components/schemas/report222"
					}
				]
			},
			"AuditDetails": {
				"enum": [
					"AUDIT_META_USER_NAME",
					"AUDIT_KEY",
					"AUDIT_BROWSER",
					"AUDIT_CHANGE_DATE",
					"AUDIT_CHANGE_TYPE",
					"AUDIT_USER",
					"AUDIT_META_OBJECT",
					"AUDIT_META_OBJECT_KEY",
					"AUDIT_META_OBJECT_PROP",
					"AUDIT_NEW_VALUE",
					"AUDIT_OLD_VALUE"
				],
				"type": "string"
			},
			"BookingData": {
				"enum": [
					"AUDIT_BOOKING_CREATED_DATE",
					"AUDIT_BOOKING_KEY",
					"AUDIT_BOOKING_CONT_KEY",
					"AUDIT_BOOKING_FULLNAME",
					"_BOOKING_SESSION_ID",
					"AUDIT_BOOKING_START_DATE"
				],
				"type": "string"
			},
			"EventDetails": {
				"enum": [
					"AUDIT_EVENT_ID",
					"AUDIT_EVENT_NAME"
				],
				"type": "string"
			},
			"SessionDetails": {
				"enum": [
					"AUDIT_SESSION_ID",
					"AUDIT_BOOKING_SESSION_NAME"
				],
				"type": "string"
			},
			"AuditBooking": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/AuditDetails"
					},
					{
						"$ref": "#/components/schemas/BookingData"
					},
					{
						"$ref": "#/components/schemas/EventDetails"
					},
					{
						"$ref": "#/components/schemas/SessionDetails"
					}
				]
			},
			"Meta": {
				"enum": [
					"AL_CHANGE_DATE",
					"AL_CHANGE_DATE_TIME",
					"AL_META_CHANGE_BROWSER",
					"AL_CHANGE_TYPE",
					"AL_META_OBJECT_KEY",
					"AL_META_OBJECT",
					"AL_USER"
				],
				"type": "string"
			},
			"MetaProp": {
				"enum": [
					"AL_META_PROP",
					"AL_NEW_VALUE",
					"AL_OLD_VALUE"
				],
				"type": "string"
			},
			"AuditLog": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/Meta"
					},
					{
						"$ref": "#/components/schemas/MetaProp"
					}
				]
			},
			"BookingSummary": {
				"enum": [
					"ASA_SESSIONS_WORKED",
					"ASA_FIRST_BOOKING",
					"ASA_TOTAL_HOURS_WORKED",
					"ASA_BOOKING_BOOKED",
					"ASA_TOTAL_PAY"
				],
				"type": "string"
			},
			"LoginSummary": {
				"enum": [
					"ASA_NO_LOGINS"
				],
				"type": "string"
			},
			"StaffAddress": {
				"enum": [
					"ASA_SITE_LOC_3",
					"ASA_SITE_COUNTY",
					"ASA_SITE_POSTCODE",
					"ASA_SITE_BUILD_1",
					"ASA_SITE_BUILD_2",
					"ASA_SITE_ID",
					"ASA_SITE_SUMM",
					"ASA_SITE_THORO_1",
					"ASA_SITE_TELE",
					"ASA_SITE_LOC_1",
					"ASA_SITE_LOC_2"
				],
				"type": "string"
			},
			"StaffAvailability": {
				"enum": [
					"ASA_AVAIL_EXTENDED_DESC",
					"ASA_AVAIL_CALANDER_DATE",
					"ASA_AVAIL_BOOL"
				],
				"type": "string"
			},
			"StaffEmployeeData": {
				"enum": [
					"ASA_EMP_BANK_ACC_NO",
					"ASA_EMP_ACTIVE",
					"ASA_EMP_ADDRESS_LAST_CHANGED",
					"ASA_PROMOTE_DATE",
					"ASA_EMP_RETIRED_DATE",
					"ASA_EMAIL_LAST_CHANGED",
					"ASA_EMAIL_LAST_CHANGED",
					"ASA_EMP_NI",
					"ASA_EMP_PASSPORT_COUNTRY",
					"ASA_EMP_PASSPORT_EXPIRY",
					"ASA_EMP_PASSPORT_NUMBER",
					"ASA_EMP_PAYROLL_NO",
					"ASA_EMP_ROLL_REF_NO",
					"ASA_EMP_BANK_SORT",
					"ASA_EMP_DESC"
				],
				"type": "string"
			},
			"StaffOrganisationData": {
				"enum": [
					"ASA_ORG_ALPHA",
					"ASA_ORG_ACC_MGR",
					"ASA_ORG_ID",
					"ASA_ORG_ID",
					"ASA_ORG_SOURCE",
					"ASA_ORG_TYPE"
				],
				"type": "string"
			},
			"StaffStatus": {
				"enum": [
					"ES_CONTACT_STATUS",
					"ASA_JOB_APP_STATUS"
				],
				"type": "string"
			},
			"StaffDetails": {
				"enum": [
					"ASA_APPLIED_DATE",
					"ASA_CONT_DOB",
					"ASA_CONTACT_ID",
					"ASA_CONT_FEATURES",
					"ASA_CONT_EMAIL",
					"ASA_CONT_1STNAME",
					"ASA_CONT_LASTNAME",
					"ASA_FULL_NAME",
					"ASA_FULL_NAME",
					"ASA_CONT_JOB_TITLE",
					"ASA_FULL_NAME_LAST_FIRST",
					"ASA_CONT_MIDDLENAME",
					"ASA_CONTACT_MOBILE",
					"ASA_CONT_PRIVATE_NOTES",
					"ASA_CONT_SALUTATION"
				],
				"type": "string"
			},
			"Lookups": {
				"properties": {
					"type": {
						"type": "string",
						"enum": [
							"AdminContactProfiles",
							"ContactProfiles",
							"OrganisationLookups",
							"SiteInformation"
						]
					},
					"value": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"type",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Common": {
				"$ref": "#/components/schemas/Lookups"
			},
			"AuditStaffActivity": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/BookingSummary"
					},
					{
						"$ref": "#/components/schemas/LoginSummary"
					},
					{
						"$ref": "#/components/schemas/StaffAddress"
					},
					{
						"$ref": "#/components/schemas/StaffAvailability"
					},
					{
						"$ref": "#/components/schemas/StaffEmployeeData"
					},
					{
						"$ref": "#/components/schemas/StaffOrganisationData"
					},
					{
						"$ref": "#/components/schemas/StaffStatus"
					},
					{
						"$ref": "#/components/schemas/StaffDetails"
					},
					{
						"$ref": "#/components/schemas/Common"
					}
				]
			},
			"CallBacks": {
				"enum": [
					"OSC_CALL_LINK",
					"OSC_CALL_CALLBACK_NO",
					"OSC_CALL_OPEN",
					"OSC_CALL_REASON",
					"OSC_CALL_TIME",
					"OSC_CALL_SETUP_BY",
					"OSC_CALL_FOR_USER",
					"OSC_CALL_CLOSED",
					"OSC_CALL_PRIORTY",
					"OSC_CALL_RESULT",
					"OSC_CALL_RESULT_DESC",
					"OSC_CALL_TYPE"
				],
				"type": "string"
			},
			"ContactDetails": {
				"enum": [
					"OSC_APPLIED_DATE",
					"OSC_ORG_LOAD_DATA",
					"OSC_CONT_DOB",
					"OSC_CONTACT_ID",
					"OSC_CONT_TELEPHONE",
					"OSC_CONT_FEATURES",
					"OSC_CONT_EMAIL",
					"OSC_CONT_1STNAME",
					"OSC_FULL_NAME",
					"OSC_CONT_LASTNAME",
					"OSC_FULL_NAME_LAST_FIRST",
					"OSC_CONT_MIDDLENAME",
					"OSC_CONTACT_MOBILE",
					"OSC_CONT_PRIVATE_NOTES",
					"OSC_CONT_SALUTATION"
				],
				"type": "string"
			},
			"ContactNotes": {
				"enum": [
					"OSC_CONTACT_NOTE1",
					"OSC_CONTACT_NOTE_DATE"
				],
				"type": "string"
			},
			"ContactStatus": {
				"enum": [
					"ES_CONTACT_STATUS"
				],
				"type": "string"
			},
			"Files": {
				"enum": [
					"OSC_FA_DATE_ADDED",
					"OSC_FA_ATTACH_DESC",
					"OSC_FA_URL",
					"OSC_FA_ATTACH_NAME",
					"OSC_FA_FILE_OBSOLETE",
					"OSC_FACT_CUST_TYPE",
					"OSC_FA_NUMBER_OF_FILES"
				],
				"type": "string"
			},
			"TalentDate": {
				"enum": [
					"OSC_TH_TALENT_VALUE"
				],
				"type": "string"
			},
			"CompaniesSitesContracts": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/CallBacks"
					},
					{
						"$ref": "#/components/schemas/ContactDetails"
					},
					{
						"$ref": "#/components/schemas/ContactNotes"
					},
					{
						"$ref": "#/components/schemas/ContactStatus"
					},
					{
						"$ref": "#/components/schemas/Files"
					},
					{
						"$ref": "#/components/schemas/TalentDate"
					},
					{
						"$ref": "#/components/schemas/Common"
					}
				]
			},
			"EventAddress": {
				"enum": [
					"ES_EVENT_ADDRESS",
					"ES_EVENT_VENUE"
				],
				"type": "string"
			},
			"Client": {
				"enum": [
					"ES_CLIENT_ACCOUNT_MANAGER",
					"ES_ORG_ALPHA",
					"ES_ORG_TYPE",
					"ES_EVENT_ORG"
				],
				"type": "string"
			},
			"ThirdPartyContact": {
				"enum": [
					"ES_EV3P_CONTACT_FULLNAME"
				],
				"type": "string"
			},
			"EventCustomFields": {
				"enum": [
					"ES_EVENT_CUSTOM_FIELD_2",
					"ES_EVENT_CUSTOM_FIELD_3",
					"ES_EVENT_CUSTOM_FIELD_4",
					"ES_EVENT_CUSTOM_FIELD_5",
					"ES_EVENT_CUSTOM_FIELD_1"
				],
				"type": "string"
			},
			"EventInvoice": {
				"enum": [
					"ES_EVENT_INVOICE_AMOUNT",
					"ES_EVENT_INVOICE_DATE",
					"ES_EVENT_INVOICE_NO",
					"ES_EVENT_INVOICE_PAID",
					"ES_EVENT_INVOICE_PAID_AMOUNT",
					"ES_EVENT_INVOICE_PAID_DATE"
				],
				"type": "string"
			},
			"EventSessionBooking": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/EventDetails"
					},
					{
						"$ref": "#/components/schemas/EventAddress"
					},
					{
						"$ref": "#/components/schemas/Client"
					},
					{
						"$ref": "#/components/schemas/ThirdPartyContact"
					},
					{
						"$ref": "#/components/schemas/EventCustomFields"
					},
					{
						"$ref": "#/components/schemas/EventInvoice"
					},
					{
						"$ref": "#/components/schemas/SessionDetails"
					}
				]
			},
			"QuestionnaireDetails": {
				"enum": [
					"QS_CONT_QS_NAME",
					"QS_CONT_QS_ID",
					"QS_CONT_QS_LINK",
					"QS_CONT_RESPONSE_DATE",
					"QS_CONT_QS_STATUS_ID",
					"QS_CONT_WEIGTHING",
					"QS_CONT_QS_STATUS",
					"QS_RESPONSE_ATTEMPTS",
					"QS_ATTEMPTS_REMAINING",
					"QS_RESPONSE_LOCKED",
					"QS_MAX_ATTEMPTS",
					"QS_UNLOCK_RESET_LINK",
					"QS_CONT_META_OBJECT_TYPE"
				],
				"type": "string"
			},
			"QuestionnaireStatus": {
				"enum": [
					"QS_CONT_NO_SUBMITTED_BY_QS",
					"QS_CONT_RESULTS_BY_QUESTIONARE_COMPLETED"
				],
				"type": "string"
			},
			"QuestionnaireProfiles": {
				"enum": [
					"QS_CONT_META_OBJECT_TYPE"
				],
				"type": "string"
			},
			"QuestionnaireResponse": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/QuestionnaireDetails"
					},
					{
						"$ref": "#/components/schemas/QuestionnaireStatus"
					},
					{
						"$ref": "#/components/schemas/QuestionnaireProfiles"
					}
				]
			},
			"filter": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/stelify"
					},
					{
						"$ref": "#/components/schemas/AuditBooking"
					},
					{
						"$ref": "#/components/schemas/AuditLog"
					},
					{
						"$ref": "#/components/schemas/AuditStaffActivity"
					},
					{
						"$ref": "#/components/schemas/CompaniesSitesContracts"
					},
					{
						"$ref": "#/components/schemas/EventSessionBooking"
					},
					{
						"$ref": "#/components/schemas/QuestionnaireResponse"
					}
				]
			},
			"rptFilters": {
				"properties": {
					"filter": {
						"$ref": "#/components/schemas/filter"
					},
					"operator": {
						"type": "string"
					},
					"values": {
						"type": "string"
					}
				},
				"required": [
					"filter",
					"operator",
					"values"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "@eventiva/axlr8",
		"version": "1.0.0",
		"contact": {
			"email": "support@resnovas.com",
			"name": "Resnovas",
			"url": "https://github.com/Resnovas/Eventiva/discussions"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/axlr8/report/{rptId}": {
			"post": {
				"operationId": "GetData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/reportResult"
								},
								"examples": {
									"Example 1": {
										"value": {
											"result": true,
											"success": [
												{
													"Account Number": " ",
													"Client:": "Example",
													"Venue:": "Example",
													"Session ID": "0000000",
													"Sess Id": "00000000",
													"Customer ID": "null",
													"No. Required": "1",
													"Number still required": "1",
													"No. DNS": "0",
													"Shift Date": "27/07/9202 09:00",
													"Start Date/Time": "null",
													"End Date/Time": "null",
													"Booking Id": "null",
													"Booking Status": "No Booking",
													"No. Confirmed": "null",
													"Staff Member Confirmed": 0,
													"Attendee Status": "No Booking",
													"Declined Status": "Not Set",
													"DNS": 0,
													"DNS Reason": "null",
													"Hour Rate": "null",
													"Shift Rate": "null",
													"Staff Rate": "null",
													"Staff Shift Rate": "null",
													"Client Hourly Rate": "null",
													"Full Name": "null",
													"NI number": "null",
													"Email": "null",
													"Operations Manager": "null",
													"Strikes": "0"
												}
											],
											"error": null,
											"debug": {
												"rptId": 1000,
												"rptFilters": [
													{
														"filter": "ES_SESSION_NUMBER_STAFF_STILL_REQUIRED",
														"operator": ">",
														"values": "0"
													},
													{
														"filter": "ES_SESSION_START_DATE",
														"operator": ">",
														"values": "09/03/2022"
													}
												]
											}
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/reportResult"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "rptId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "url",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "api_user",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "api_user_pw",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/rptFilters"
								},
								"type": "array"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "https://localhost:3000/"
		}
	]
}